refactor:  App, index, GlobalStyles 파일 ts 파일로 변경
#1344 #1372
